requirements.txt
selenium
chromedriver-binary==78.0.3904.70.0
pyserial
